pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rsu-tax"
version = "0.1.0"
description = "RSU capital gains calculator for German tax reporting (Abgeltungssteuer)"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "jinja2>=3.1",
    "python-multipart>=0.0.12",
    "httpx>=0.27",
    "pydantic>=2.0",
    "fpdf2>=2.7",
]

[project.scripts]
rsu-tax = "rsu_tax.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[tool.hatch.build.targets.wheel]
packages = ["src/rsu_tax"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
